# packages/contracts/openapi/api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/authors": {
      "get": {
        "operationId": "AuthorsAdminController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      },
      "post": {
        "operationId": "AuthorsAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/authors/{id}": {
      "get": {
        "operationId": "AuthorsAdminController_getAuthor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDto"
                }
              }
            }
          }
        },
        "summary": "One author profile with links, photo and counts",
        "tags": [
          "admin-blog"
        ]
      },
      "patch": {
        "operationId": "AuthorsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/authors/{id}/activate": {
      "post": {
        "operationId": "AuthorsAdminController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogTermStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/authors/{id}/deactivate": {
      "post": {
        "operationId": "AuthorsAdminController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogTermStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/blog-categories": {
      "get": {
        "operationId": "BlogCategoriesAdminController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogTermDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      },
      "post": {
        "operationId": "BlogCategoriesAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogTermInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/blog-categories/{id}": {
      "patch": {
        "operationId": "BlogCategoriesAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogTermDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/blog-categories/{id}/activate": {
      "post": {
        "operationId": "BlogCategoriesAdminController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogTermStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/blog-categories/{id}/deactivate": {
      "post": {
        "operationId": "BlogCategoriesAdminController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogTermStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/blog-tags": {
      "get": {
        "operationId": "BlogTagsAdminController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogTermDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      },
      "post": {
        "operationId": "BlogTagsAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogTermInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/blog-tags/{id}": {
      "patch": {
        "operationId": "BlogTagsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogTermDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/blog-tags/{id}/activate": {
      "post": {
        "operationId": "BlogTagsAdminController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogTermStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/blog-tags/{id}/deactivate": {
      "post": {
        "operationId": "BlogTagsAdminController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogTermStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/posts": {
      "get": {
        "operationId": "PostsAdminController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "scheduled",
                "published",
                "archived"
              ]
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Matches title and slug",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "updatedAt",
              "type": "string",
              "enum": [
                "updatedAt",
                "publishedAt",
                "scheduledAt",
                "title"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List articles (status, category, tag, author, q, sort, order, page, pageSize)",
        "tags": [
          "admin-blog"
        ]
      },
      "post": {
        "operationId": "PostsAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/posts/{id}": {
      "get": {
        "operationId": "PostsAdminController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      },
      "patch": {
        "operationId": "PostsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/posts/{id}/preview": {
      "get": {
        "operationId": "PostsAdminController_preview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPreviewDto"
                }
              }
            }
          }
        },
        "summary": "Sanitised preview for authorised admins only; never public and never indexable (SRS BLOG 003)",
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/posts/{id}/revisions": {
      "get": {
        "operationId": "PostsAdminController_revisions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostRevisionDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/posts/{id}/slug": {
      "post": {
        "operationId": "PostsAdminController_changeSlug",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSlugDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "summary": "Change the public slug; a published article keeps a 301 from the old path (SRS SEO 004)",
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/posts/{id}/publish": {
      "post": {
        "operationId": "PostsAdminController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/posts/{id}/schedule": {
      "post": {
        "operationId": "PostsAdminController_schedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "summary": "Schedule publication at a UTC instant chosen as Melbourne time",
        "tags": [
          "admin-blog"
        ]
      }
    },
    "/api/v1/admin/posts/{id}/unpublish": {
      "post": {
        "operationId": "PostsAdminController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-blog"
        ]
      }
    },
...
  },
  "info": {
    "title": "Melbourne Sphere API",
    "description": "REST API for the Melbourne Sphere directory and blog. Admin routes require the ms_admin_session cookie.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ms_admin_session"
      }
    },
    "schemas": {
...
      "BlogTermStateDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "expectedVersion"
        ]
      },
      "BlogTermDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "landingContent": {
            "type": "string",
            "nullable": true,
            "description": "Sanitised HTML rendered on the landing page"
          },
          "active": {
            "type": "boolean"
          },
          "postCount": {
            "type": "number"
          },
          "version": {
            "type": "number"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "landingContent",
          "active",
          "postCount",
          "version",
          "updatedAt"
        ]
      },
      "BlogTermInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "slug": {
            "type": "string",
            "maxLength": 100
          },
          "landingContent": {
            "type": "string",
            "nullable": true,
            "maxLength": 5000,
            "description": "Editorial landing content (Markdown); a tag without it is noindex (SRS BLOG 005)"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}